auth/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.0
boto3==1.28.10
botocore==1.31.10
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.6
confluent-kafka==2.2.0
cryptography==41.0.2
cytoolz==0.12.2
defusedxml==0.7.1
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-js-asset==2.1.0
django-redis==5.3.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-api-response==0.0.30
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
idna==3.4
jmespath==1.0.1
jsonpickle==3.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
parsimonious==0.9.0
pika==1.3.2
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
pyunormalize==15.0.0
redis==4.6.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.5.0
text-unidecode==1.3
toolz==0.12.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.1
web3==6.6.1
websockets==11.0.3
whitenoise==6.5.0
yarl==1.9.2
